test(grid): split non-grid suites out of GridLayoutTests
GridLayoutTests.swift had grown past the 350-line ceiling (517) after the
#171 cap/auto-size tests landed. Move the three non-grid suites it
carried — the layout dispatcher / monocle-floating basics, gap geometry,
and the float-rule / retile filter — into LayoutMiscTests.swift, where
they belong by topic anyway. Both files now sit well under the ceiling;
no tests change (1017 still pass). Addresses the branch review's one
actionable finding (AGENTS.md §5, split suites early).
